Factor out signal client interface.
