+++ packages/core/mesh/messaging/src/signal-manager/signal-manager.ts
@@ -6,20 +6,20 @@ import { type Event } from '@dxos/async';
 import { type PublicKey } from '@dxos/keys';
 import { type SwarmEvent } from '@dxos/protocols/proto/dxos/mesh/signal';
 
-import { type CommandTrace, type SignalStatus } from '../signal-client';
-import { type Message, type SignalMethods } from '../signal-methods';
+import { type CommandTrace } from '../signal-client';
+import { type Message, type SignalMethods, type SignalStatus } from '../signal-methods';
 
 /**
- *
+ * Manages a collection of signaling clients.
  */
 export interface SignalManager extends SignalMethods {
+  open(): Promise<void>;
+  close(): Promise<void>;
+
+  getStatus(): SignalStatus[];
+
   statusChanged: Event<SignalStatus[]>;
   commandTrace: Event<CommandTrace>;
   swarmEvent: Event<{ topic: PublicKey; swarmEvent: SwarmEvent }>;
   onMessage: Event<Message>;
-
-  getStatus(): SignalStatus[];
-
-  open(): Promise<void>;
-  close(): Promise<void>;
 }
