+++ packages/core/mesh/messaging/src/signal-methods.ts
@@ -4,6 +4,7 @@
 
 import { type Any } from '@dxos/codec-protobuf';
 import { type PublicKey } from '@dxos/keys';
+import { type SignalState } from '@dxos/protocols/proto/dxos/mesh/signal';
 
 export interface Message {
   author: PublicKey;
@@ -11,6 +12,18 @@ export interface Message {
   payload: Any;
 }
 
+export type SignalStatus = {
+  host: string;
+  state: SignalState;
+  error?: string;
+  reconnectIn: number;
+  connectionStarted: Date;
+  lastStateChange: Date;
+};
+
+/**
+ * Message routing interface.
+ */
 export interface SignalMethods {
   /**
    * Join topic on signal network, to be discoverable by other peers.
@@ -36,4 +49,13 @@ export interface SignalMethods {
    * Stop receiving messages from peer.
    */
   unsubscribeMessages: (peerId: PublicKey) => Promise<void>;
+}
+
+/**
+ * Signaling client.
+ */
+export interface SignalClientMethods extends SignalMethods {
+  open(): Promise<void>;
+  close(): Promise<void>;
+  getStatus(): SignalStatus;
 }
